move models to models.py
