+++ app/models.py
@@ -7,3 +7,11 @@ class User(db.Model):
     lastname = db.Column(db.String(128), nullable=False)
     email = db.Column(db.String(128), unique=True, nullable=False)
     password = db.Column(db.String(255), nullable=False)
+
+
+class Transaction(db.Model):
+    __tablename__ = "transactions"
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    date = db.Column(db.Date(), nullable=False)
+    description = db.Column(db.String(255), nullable=False)
+    amount = db.Column(db.Integer(), nullable=False)
